Merges files into one
